handle occasional failure to remove channel from watchlist
When testing tclkit sometimes a closed channel was being left in the
list of mk channels. This then causes a crash as we attempt to access
the deleted memory. Solved by passing the MkWorkspace reference in the
channel instance data and calling Tcl_DeleteEvents before we unhook the
channel from the list.
Added a test using simultanous fcopy from metakit.

Signed-off-by: Pat Thoyts <patthoyts@users.sourceforge.net>